+++ RestNET5/Services/IPersonService.cs
@@ -6,10 +6,10 @@ namespace RestNET5.Services
     public interface IPersonService
     {
         Person Create(Person person);
-        Person FindByID(double id);
+        Person FindByID(long id);
         List<Person> FindAll();
 
         Person Update(Person person);
-        void Delete(double id);
+        void Delete(long id);
     }
 }
